handle invitations where group is not available anymore (ie: drop them)
continue destroy when membership.group is nil

+++ app/models/membership.rb
@@ -88,6 +88,7 @@ class Membership < ActiveRecord::Base
   end
 
   def destroy_subgroup_memberships
+    return if group.nil? #necessary if group is missing (as in case of production data)
     group.subgroups.each do |subgroup|
       membership = subgroup.memberships.find_by_user_id(user.id)
       membership.destroy if membership
@@ -95,6 +96,7 @@ class Membership < ActiveRecord::Base
   end
 
   def remove_open_votes
+    return if group.nil? #necessary if group is missing (as in case of production data)
     discussions = Queries::VisibleDiscussions.for(group, user)
     discussions.with_current_motions_user_has_voted_on.each do |discussion|
       votes = discussion.current_motion.votes.where(:user_id => user.id)
